Add impact amount for fraction of 1
The impact calculator will now show a % value of a donations effectiveness, if the impact amount is less than 1. For example, if it costs $5 to provide 1 bed net for malaria, and someone donates $4, it will show that their donation will provide 80% of what's needed for a single bednet.

Organisation costs have been changed to USD, because in general, organisations provide their figures in USD. This means that to convert to AUD only the usdToAudExchangeRate value needs to be changed periodically.